+++ src/theory/booleans/proof_circuit_propagator.h
@@ -23,42 +23,10 @@
 
 #include "expr/node.h"
 #include "expr/proof_node.h"
-#include "theory/booleans/circuit_propagator.h"
 
 namespace CVC4 {
 namespace theory {
 namespace booleans {
-
-namespace {
-
-/** Shorthand to create a Node from a constant number */
-template <typename T>
-Node mkRat(T val)
-{
-  return NodeManager::currentNM()->mkConst<Rational>(val);
-}
-
-/**
- * Collect all children from parent except for one particular child (the
- * "holdout"). The holdout is given as iterator, and should be an iterator into
- * the [parent.begin(),parent.end()] range.
- */
-inline std::vector<Node> collectButHoldout(TNode parent,
-                                           TNode::iterator holdout)
-{
-  std::vector<Node> lits;
-  for (TNode::iterator i = parent.begin(), i_end = parent.end(); i != i_end;
-       ++i)
-  {
-    if (i != holdout)
-    {
-      lits.emplace_back(*i);
-    }
-  }
-  return lits;
-}
-
-}  // namespace
 
 /**
  * Base class for for CircuitPropagatorProofs.
@@ -71,48 +39,18 @@ struct ProofCircuitPropagator
   ProofNodeManager* d_pnm;
 
   /** Shorthand to check whether proof generation is disabled */
-  bool disabled() const { return d_pnm == nullptr; }
+  bool disabled() const;
 
   /** Construct proof by assuming the given node */
-  std::shared_ptr<ProofNode> mkProof(Node n) { return d_pnm->mkAssume(n); }
+  std::shared_ptr<ProofNode> mkProof(Node n);
   /** Construct proof using the given rule, children and args */
   std::shared_ptr<ProofNode> mkProof(
       PfRule rule,
       const std::vector<std::shared_ptr<ProofNode>>& children,
-      const std::vector<Node>& args = {})
-  {
-    if (Trace.isOn("circuit-prop"))
-    {
-      std::stringstream ss;
-      ss << "Constructing (" << rule;
-      for (const auto& c : children)
-      {
-        ss << " " << *c;
-      }
-      if (!args.empty())
-      {
-        ss << " :args";
-        for (const auto& a : args)
-        {
-          ss << " " << a;
-        }
-      }
-      ss << ")";
-      Trace("circuit-prop") << ss.str() << std::endl;
-    }
-    return d_pnm->mkNode(rule, children, args);
-  }
+      const std::vector<Node>& args = {});
   /** Apply CONTRA rule. Takes care of switching a and b if necessary */
   std::shared_ptr<ProofNode> mkContra(const std::shared_ptr<ProofNode>& a,
-                                      const std::shared_ptr<ProofNode>& b)
-  {
-    if (a->getResult().notNode() == b->getResult())
-    {
-      return mkProof(PfRule::CONTRA, {a, b});
-    }
-    Assert(a->getResult() == b->getResult().notNode());
-    return mkProof(PfRule::CONTRA, {b, a});
-  }
+                                      const std::shared_ptr<ProofNode>& b);
   /**
    * Apply CHAIN_RESOLUTION rule.
    * Constructs the args from the given literals and polarities (called ids in
@@ -121,88 +59,24 @@ struct ProofCircuitPropagator
    */
   std::shared_ptr<ProofNode> mkResolution(std::shared_ptr<ProofNode> clause,
                                           const std::vector<Node>& lits,
-                                          const std::vector<bool>& polarity)
-  {
-    auto* nm = NodeManager::currentNM();
-    std::vector<std::shared_ptr<ProofNode>> children = {clause};
-    std::vector<Node> args;
-    Assert(lits.size() == polarity.size());
-    for (std::size_t i = 0; i < lits.size(); ++i)
-    {
-      bool pol = polarity[i];
-      Node lit = lits[i];
-      if (polarity[i])
-      {
-        if (lits[i].getKind() == Kind::NOT)
-        {
-          lit = lit[0];
-          pol = !pol;
-          children.emplace_back(mkProof(lit));
-        }
-        else
-        {
-          children.emplace_back(mkProof(lits[i].notNode()));
-        }
-      }
-      else
-      {
-        children.emplace_back(mkProof(lits[i]));
-      }
-      args.emplace_back(nm->mkConst<bool>(pol));
-      args.emplace_back(lit);
-    }
-    return mkProof(PfRule::CHAIN_RESOLUTION, children, args);
-  }
+                                          const std::vector<bool>& polarity);
   /** Shorthand for mkResolution(clause, lits, {polarity, ...}) */
   std::shared_ptr<ProofNode> mkResolution(std::shared_ptr<ProofNode> clause,
                                           const std::vector<Node>& lits,
-                                          bool polarity)
-  {
-    return mkResolution(clause, lits, std::vector<bool>(lits.size(), polarity));
-  }
+                                          bool polarity);
   /** Apply NOT_NOT_ELIM rule if n.getResult() is a nested negation */
-  std::shared_ptr<ProofNode> mkNot(const std::shared_ptr<ProofNode>& n)
-  {
-    Node m = n->getResult();
-    if (m.getKind() == Kind::NOT && m[0].getKind() == Kind::NOT)
-    {
-      return mkProof(PfRule::NOT_NOT_ELIM, {n});
-    }
-    return n;
-  }
+  std::shared_ptr<ProofNode> mkNot(const std::shared_ptr<ProofNode>& n);
 
   /** (and true ... holdout true ...)  -->  holdout */
-  std::shared_ptr<ProofNode> andFalse(Node parent, TNode::iterator holdout)
-  {
-    if (disabled()) return nullptr;
-    return mkNot(
-        mkResolution(mkProof(PfRule::NOT_AND, {mkProof(parent.notNode())}),
-                     collectButHoldout(parent, holdout),
-                     false));
-  }
+  std::shared_ptr<ProofNode> andFalse(Node parent, TNode::iterator holdout);
 
   /** (or false ... holdout false ...)  ->  holdout */
-  std::shared_ptr<ProofNode> orTrue(Node parent, TNode::iterator holdout)
-  {
-    if (disabled()) return nullptr;
-    return mkResolution(
-        mkProof(parent), collectButHoldout(parent, holdout), true);
-  }
+  std::shared_ptr<ProofNode> orTrue(Node parent, TNode::iterator holdout);
 
   /** (=> X false)  -->  (not X) */
-  std::shared_ptr<ProofNode> impliesXFromY(Node parent)
-  {
-    if (disabled()) return nullptr;
-    return mkNot(mkResolution(
-        mkProof(PfRule::IMPLIES_ELIM, {mkProof(parent)}), {parent[1]}, {true}));
-  }
+  std::shared_ptr<ProofNode> impliesXFromY(Node parent);
   /** (=> true Y)  -->  Y */
-  std::shared_ptr<ProofNode> impliesYFromX(Node parent)
-  {
-    if (disabled()) return nullptr;
-    return mkResolution(
-        mkProof(PfRule::IMPLIES_ELIM, {mkProof(parent)}), {parent[0]}, {false});
-  }
+  std::shared_ptr<ProofNode> impliesYFromX(Node parent);
 
   /**
    * Uses (xor X Y) to derive the value of X.
@@ -211,26 +85,7 @@ struct ProofCircuitPropagator
    * (not (xor X false))  -->  (not X)
    * (not (xor X true))  -->  X
    */
-  std::shared_ptr<ProofNode> mkXorXFromY(bool negated, bool y, Node parent)
-  {
-    if (disabled()) return nullptr;
-    if (y)
-    {
-      return mkNot(mkResolution(
-          mkProof(negated ? PfRule::NOT_XOR_ELIM1 : PfRule::XOR_ELIM2,
-                  {mkProof(negated ? parent.notNode() : Node(parent))}),
-          {parent[1]},
-          {false}));
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(negated ? PfRule::NOT_XOR_ELIM2 : PfRule::XOR_ELIM1,
-                  {mkProof(negated ? parent.notNode() : Node(parent))}),
-          {parent[1]},
-          {true});
-    }
-  }
+  std::shared_ptr<ProofNode> mkXorXFromY(bool negated, bool y, Node parent);
   /**
    * Uses (xor X Y) to derive the value of Y.
    * (xor false Y)  -->  Y
@@ -238,26 +93,7 @@ struct ProofCircuitPropagator
    * (not (xor false Y))  -->  (not Y)
    * (not (xor true Y))  -->  Y
    */
-  std::shared_ptr<ProofNode> mkXorYFromX(bool negated, bool x, Node parent)
-  {
-    if (disabled()) return nullptr;
-    if (x)
-    {
-      return mkResolution(
-          mkProof(negated ? PfRule::NOT_XOR_ELIM2 : PfRule::XOR_ELIM1,
-                  {mkProof(negated ? parent.notNode() : Node(parent))}),
-          {parent[0]},
-          {false});
-    }
-    else
-    {
-      return mkNot(mkResolution(
-          mkProof(negated ? PfRule::NOT_XOR_ELIM1 : PfRule::XOR_ELIM2,
-                  {mkProof(negated ? parent.notNode() : Node(parent))}),
-          {parent[0]},
-          {true}));
-    }
-  }
+  std::shared_ptr<ProofNode> mkXorYFromX(bool negated, bool x, Node parent);
 };
 
 /** Proof generator for backward propagation */
@@ -268,36 +104,16 @@ struct ProofCircuitPropagatorBackward : public ProofCircuitPropagator
 
   ProofCircuitPropagatorBackward(ProofNodeManager* pnm,
                                  TNode parent,
-                                 bool parentAssignment)
-      : ProofCircuitPropagator{pnm},
-        d_parent(parent),
-        d_parentAssignment(parentAssignment)
-  {
-  }
+                                 bool parentAssignment);
 
   /** and true  -->  child is true */
-  std::shared_ptr<ProofNode> andTrue(TNode::iterator i)
-  {
-    if (disabled()) return nullptr;
-    return mkProof(
-        PfRule::AND_ELIM, {mkProof(d_parent)}, {mkRat(i - d_parent.begin())});
-  }
+  std::shared_ptr<ProofNode> andTrue(TNode::iterator i);
 
   /** or false  -->  child is false */
-  std::shared_ptr<ProofNode> orFalse(TNode::iterator i)
-  {
-    if (disabled()) return nullptr;
-    return mkNot(mkProof(PfRule::NOT_OR_ELIM,
-                         {mkProof(d_parent.notNode())},
-                         {mkRat(i - d_parent.begin())}));
-  }
+  std::shared_ptr<ProofNode> orFalse(TNode::iterator i);
 
   /** (not x) is true  -->  x is false (and vice versa) */
-  std::shared_ptr<ProofNode> Not()
-  {
-    if (disabled()) return nullptr;
-    return mkNot(mkProof(d_parentAssignment ? d_parent : d_parent[0]));
-  }
+  std::shared_ptr<ProofNode> Not();
 
   /**
    * Propagate on ite with evaluate condition
@@ -306,134 +122,27 @@ struct ProofCircuitPropagatorBackward : public ProofCircuitPropagator
    * (ite false t e)  -->  e
    * (not (ite false t e))  -->  (not e)
    */
-  std::shared_ptr<ProofNode> iteC(bool c)
-  {
-    if (disabled()) return nullptr;
-    if (d_parentAssignment)
-    {
-      return mkResolution(mkProof(c ? PfRule::ITE_ELIM1 : PfRule::ITE_ELIM2,
-                                  {mkProof(d_parent)}),
-                          {d_parent[0]},
-                          {!c});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(c ? PfRule::NOT_ITE_ELIM1 : PfRule::NOT_ITE_ELIM2,
-                  {mkProof(d_parent.notNode())}),
-          {d_parent[0]},
-          {c});
-    }
-  }
+  std::shared_ptr<ProofNode> iteC(bool c);
   /**
    * For (ite c t e), we can derive the value for c
    * c = 1: c = true
    * c = 0: c = false
    */
-  std::shared_ptr<ProofNode> iteIsCase(unsigned c)
-  {
-    if (disabled()) return nullptr;
-    if (d_parentAssignment)
-    {
-      return mkResolution(mkProof(PfRule::ITE_ELIM2, {mkProof(d_parent)}),
-                          {d_parent[c + 1]}, {false});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_ITE_ELIM2, {mkProof(d_parent.notNode())}),
-          {d_parent[c + 1]}, {false});
-    }
-  }
+  std::shared_ptr<ProofNode> iteIsCase(unsigned c);
 
   /** Derive X from (= X Y) */
-  std::shared_ptr<ProofNode> eqXFromY(bool y)
-  {
-    if (disabled()) return nullptr;
-    if (y)
-    {
-      return mkProof(
-          PfRule::EQ_RESOLVE,
-          {mkProof(d_parent[1]), mkProof(PfRule::SYMM, {mkProof(d_parent)})});
-    }
-    else
-    {
-      return mkNot(
-          mkResolution(mkProof(PfRule::EQUIV_ELIM1, {mkProof(d_parent)}),
-                       {d_parent[1]},
-                       {true}));
-    }
-  }
+  std::shared_ptr<ProofNode> eqXFromY(bool y);
   /** Derive Y from (= X Y) */
-  std::shared_ptr<ProofNode> eqYFromX(bool x)
-  {
-    if (disabled()) return nullptr;
-    if (x)
-    {
-      return mkProof(PfRule::EQ_RESOLVE,
-                     {mkProof(d_parent[0]), mkProof(d_parent)});
-    }
-    else
-    {
-      return mkNot(
-          mkResolution(mkProof(PfRule::EQUIV_ELIM2, {mkProof(d_parent)}),
-                       {d_parent[0]},
-                       {true}));
-    }
-  }
+  std::shared_ptr<ProofNode> eqYFromX(bool x);
   /** Derive X from (not (= X Y)) */
-  std::shared_ptr<ProofNode> neqXFromY(bool y)
-  {
-    if (disabled()) return nullptr;
-    if (y)
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM2, {mkProof(d_parent.notNode())}),
-          {d_parent[1]},
-          {false});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM1, {mkProof(d_parent.notNode())}),
-          {d_parent[1]},
-          {true});
-    }
-  }
+  std::shared_ptr<ProofNode> neqXFromY(bool y);
   /** Derive Y from (not (= X Y)) */
-  std::shared_ptr<ProofNode> neqYFromX(bool x)
-  {
-    if (disabled()) return nullptr;
-    if (x)
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM2, {mkProof(d_parent.notNode())}),
-          {d_parent[0]},
-          {false});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM1, {mkProof(d_parent.notNode())}),
-          {d_parent[0]},
-          {true});
-    }
-  }
+  std::shared_ptr<ProofNode> neqYFromX(bool x);
 
   /** (not (=> X Y))  -->  X */
-  std::shared_ptr<ProofNode> impliesNegX()
-  {
-    if (disabled()) return nullptr;
-    return mkNot(
-        mkProof(PfRule::NOT_IMPLIES_ELIM1, {mkProof(d_parent.notNode())}));
-  }
+  std::shared_ptr<ProofNode> impliesNegX();
   /** (not (=> X Y))  -->  (not Y) */
-  std::shared_ptr<ProofNode> impliesNegY()
-  {
-    if (disabled()) return nullptr;
-    return mkNot(
-        mkProof(PfRule::NOT_IMPLIES_ELIM2, {mkProof(d_parent.notNode())}));
-  }
+  std::shared_ptr<ProofNode> impliesNegY();
 };
 
 /** Proof generator for forward propagation */
@@ -446,244 +155,43 @@ struct ProofCircuitPropagatorForward : public ProofCircuitPropagator
   ProofCircuitPropagatorForward(ProofNodeManager* pnm,
                                 Node child,
                                 bool childAssignment,
-                                Node parent)
-      : ProofCircuitPropagator{pnm},
-        d_child(child),
-        d_childAssignment(childAssignment),
-        d_parent(parent)
-  {
-  }
+                                Node parent);
 
   /** All children are true  -->  and is true */
-  std::shared_ptr<ProofNode> andAllTrue()
-  {
-    if (disabled()) return nullptr;
-    std::vector<std::shared_ptr<ProofNode>> children;
-    for (const auto& child : d_parent)
-    {
-      children.emplace_back(mkProof(child));
-    }
-    return mkProof(PfRule::AND_INTRO, children);
-  }
+  std::shared_ptr<ProofNode> andAllTrue();
   /** One child is false  -->  and is false */
-  std::shared_ptr<ProofNode> andOneFalse()
-  {
-    if (disabled()) return nullptr;
-    auto it = std::find(d_parent.begin(), d_parent.end(), d_child);
-    return mkProof(PfRule::SCOPE,
-                   {mkContra(mkProof(PfRule::AND_ELIM,
-                                     {mkProof(d_parent)},
-                                     {mkRat(it - d_parent.begin())}),
-                             mkProof(d_child.notNode()))},
-                   {d_parent});
-  }
+  std::shared_ptr<ProofNode> andOneFalse();
 
   /** One child is true  -->  or is true */
-  std::shared_ptr<ProofNode> orOneTrue()
-  {
-    if (disabled()) return nullptr;
-    auto it = std::find(d_parent.begin(), d_parent.end(), d_child);
-    return mkNot(mkResolution(
-        mkProof(
-            PfRule::CNF_OR_NEG, {}, {d_parent, mkRat(it - d_parent.begin())}),
-        {d_child},
-        {false}));
-  }
+  std::shared_ptr<ProofNode> orOneTrue();
   /** or false  -->  all children are false */
-  std::shared_ptr<ProofNode> orFalse()
-  {
-    if (disabled()) return nullptr;
-    std::vector<Node> children(d_parent.begin(), d_parent.end());
-    return mkResolution(
-        mkProof(PfRule::CNF_OR_POS, {}, {d_parent}), children, true);
-  }
+  std::shared_ptr<ProofNode> orFalse();
 
   /** Evaluate (ite true X _) from X */
-  std::shared_ptr<ProofNode> iteEvalThen(bool x)
-  {
-    if (disabled()) return nullptr;
-    if (x)
-    {
-      return mkResolution(mkProof(PfRule::CNF_ITE_NEG1, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {false, false});
-    }
-    else
-    {
-      return mkResolution(mkProof(PfRule::CNF_ITE_POS1, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {false, true});
-    }
-  }
+  std::shared_ptr<ProofNode> iteEvalThen(bool x);
   /** Evaluate (ite false _ Y) from Y */
-  std::shared_ptr<ProofNode> iteEvalElse(bool y)
-  {
-    if (disabled()) return nullptr;
-    if (y)
-    {
-      return mkResolution(mkProof(PfRule::CNF_ITE_NEG2, {}, {d_parent}),
-                          {d_parent[0], d_parent[2]},
-                          {true, false});
-    }
-    else
-    {
-      return mkResolution(mkProof(PfRule::CNF_ITE_POS2, {}, {d_parent}),
-                          {d_parent[0], d_parent[2]},
-                          {true, true});
-    }
-  }
+  std::shared_ptr<ProofNode> iteEvalElse(bool y);
 
   /** x is true  -->  (not x) is false (and vice versa) */
-  std::shared_ptr<ProofNode> Not()
-  {
-    if (disabled()) return nullptr;
-    return mkNot(mkProof(d_childAssignment ? d_parent[0] : Node(d_parent)));
-  }
+  std::shared_ptr<ProofNode> Not();
 
   /** Evaluate (= X Y) from X,Y */
-  std::shared_ptr<ProofNode> eqEval()
-  {
-    if (disabled()) return nullptr;
-    if (d_childAssignment)
-    {
-      return mkResolution(mkProof(PfRule::CNF_EQUIV_NEG2, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {false, false});
-    }
-    else
-    {
-      return mkResolution(mkProof(PfRule::CNF_EQUIV_NEG1, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {true, true});
-    }
-  }
+  std::shared_ptr<ProofNode> eqEval();
   /** Derive Y from (= X Y) */
-  std::shared_ptr<ProofNode> eqYFromX()
-  {
-    if (disabled()) return nullptr;
-    Assert(d_parent[0] == d_child);
-    if (d_childAssignment)
-    {
-      return mkProof(PfRule::EQ_RESOLVE, {mkProof(d_child), mkProof(d_parent)});
-    }
-    else
-    {
-      return mkNot(
-          mkResolution(mkProof(PfRule::EQUIV_ELIM2, {mkProof(d_parent)}),
-                       {d_child},
-                       {true}));
-    }
-  }
+  std::shared_ptr<ProofNode> eqYFromX();
 
   /** Derive Y from (not (= X Y)) */
-  std::shared_ptr<ProofNode> neqYFromX()
-  {
-    if (disabled()) return nullptr;
-    Assert(d_parent[0] == d_child);
-    if (d_childAssignment)
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM2, {mkProof(d_parent.notNode())}),
-          {d_child},
-          {false});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(PfRule::NOT_EQUIV_ELIM1, {mkProof(d_parent.notNode())}),
-          {d_child},
-          {true});
-    }
-  }
+  std::shared_ptr<ProofNode> neqYFromX();
 
   /** Derive X from (= X Y) */
-  std::shared_ptr<ProofNode> eqXFromY()
-  {
-    if (disabled()) return nullptr;
-    Assert(d_parent[1] == d_child);
-    if (d_childAssignment)
-    {
-      return mkProof(
-          PfRule::EQ_RESOLVE,
-          {mkProof(d_child), mkProof(PfRule::SYMM, {mkProof(d_parent)})});
-    }
-    else
-    {
-      return mkResolution(
-          mkProof(PfRule::EQUIV_ELIM1, {mkProof(d_parent)}), {d_child}, {true});
-    }
-  }
+  std::shared_ptr<ProofNode> eqXFromY();
   /** Derive X from (not (= X Y)) */
-  std::shared_ptr<ProofNode> neqXFromY()
-  {
-    if (disabled()) return nullptr;
-    Assert(d_parent[0] == d_child);
-    if (d_childAssignment)
-    {
-      return mkResolution(mkProof(PfRule::NOT_EQUIV_ELIM2,
-                                  {mkProof(PfRule::SYMM, {mkProof(d_parent)})}),
-                          {d_child},
-                          {false});
-    }
-    else
-    {
-      return mkResolution(mkProof(PfRule::NOT_EQUIV_ELIM1,
-                                  {mkProof(PfRule::SYMM, {mkProof(d_parent)})}),
-                          {d_child},
-                          {true});
-    }
-  }
+  std::shared_ptr<ProofNode> neqXFromY();
 
   /** Evaluate (=> X Y) from X,Y */
-  std::shared_ptr<ProofNode> impliesEval(bool premise, bool conclusion)
-  {
-    if (disabled()) return nullptr;
-    if (!premise)
-    {
-      return mkResolution(mkProof(PfRule::CNF_IMPLIES_NEG1, {}, {d_parent}),
-                          {d_parent[0]},
-                          {true});
-    }
-    if (conclusion)
-    {
-      return mkResolution(mkProof(PfRule::CNF_IMPLIES_NEG2, {}, {d_parent}),
-                          {d_parent[1]},
-                          {false});
-    }
-    return mkResolution(mkProof(PfRule::CNF_IMPLIES_POS, {}, {d_parent}),
-                        {d_parent[0], d_parent[1]},
-                        {false, true});
-  }
+  std::shared_ptr<ProofNode> impliesEval(bool premise, bool conclusion);
   /** Evaluate (xor X Y) from X,Y */
-  std::shared_ptr<ProofNode> xorEval(bool x, bool y)
-  {
-    if (disabled()) return nullptr;
-    if (x && y)
-    {
-      return mkResolution(mkProof(PfRule::CNF_XOR_POS2, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {false, false});
-    }
-    else if (x && !y)
-    {
-      return mkResolution(mkProof(PfRule::CNF_XOR_NEG1, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {false, true});
-    }
-    else if (!x && y)
-    {
-      return mkResolution(mkProof(PfRule::CNF_XOR_NEG2, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {true, false});
-    }
-    else
-    {
-      Assert(!x && !y);
-      return mkResolution(mkProof(PfRule::CNF_XOR_POS1, {}, {d_parent}),
-                          {d_parent[0], d_parent[1]},
-                          {true, true});
-    }
-  }
+  std::shared_ptr<ProofNode> xorEval(bool x, bool y);
 };
 
 }  // namespace booleans
